AiService/app.py
"""
Flask API for TensorFlow Anti-Cheat Service
"""
from flask import Flask, request, jsonify
import os
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'service': 'TensorFlow Anti-Cheat Service',
        'version': '1.0.0'
    }), 200

@app.route('/validate_move', methods=['POST'])
def validate_move():
    """
    Validate player movement for cheating
    Expected JSON: {
        "player_id": "string",
        "move": {"x": float, "y": float, "z": float},
        "velocity": float,
        "timestamp": float
    }
    """
    try:
        data = request.get_json()
        
        # TODO: Implement TensorFlow model inference
        # For now, return a basic validation
        
        return jsonify({
            'valid': True,
            'confidence': 0.95,
            'message': 'Move approved'
        }), 200
    except Exception as e:
        return jsonify({
            'error': str(e)
        }), 400

@app.route('/detect_cheat', methods=['POST'])
def detect_cheat():
    """
    Detect cheating behavior using TensorFlow
    Expected JSON: {
        "player_id": "string",
        "game_data": {...}
    }
    """
    try:
        data = request.get_json()
        
        # TODO: Implement TensorFlow model inference
        
        return jsonify({
            'is_cheating': False,
            'probability': 0.02,
            'reason': 'Normal gameplay'
        }), 200
    except Exception as e:
        return jsonify({
            'error': str(e)
        }), 400

if __name__ == '__main__':
    port = int(os.getenv('FLASK_PORT', 5000))
    debug = os.getenv('FLASK_ENV') == 'development'
    app.run(host='0.0.0.0', port=port, debug=debug)
